Created Coordinate interface and renamed og coodinate to TTTTPoisition
