+++ src/edu/saddleback/cs4b/Backend/Messages/Coordinate.java
@@ -1,36 +1,9 @@
 package edu.saddleback.cs4b.Backend.Messages;
 
-public class Coordinate {
-    private int xCoord;
-    private int yCoord;
-
-    public Coordinate()
-    {
-
-    }
-    public Coordinate(int newX, int newY)
-    {
-        xCoord = newX;
-        yCoord = newY;
-    }
-
-    public int getXCoord()
-    {
-        return xCoord;
-    }
-
-    public void setXCoord(int newXCoord)
-    {
-        this.xCoord = xCoord;
-    }
-
-    public int getYCoord()
-    {
-        return yCoord;
-    }
-
-    public void setYCoord(int newYCoord)
-    {
-        this.yCoord = yCoord;
-    }
+public interface Coordinate
+{
+    public int  getXCoord();
+    public void setXCoord(int newXCoord);
+    public int  getYCoord();
+    public void setYCoord(int newYCoord);
 }
